+++ src/scala/co/actioniq/luna/dao/JdbcTypeImplicits.scala
@@ -42,7 +42,7 @@ class JdbcTypeImplicits[P <: JdbcProfile](val profile: P) {
 }
 
 object JdbcTypeImplicits {
-  val h2JdbcTypeImplicits = new JdbcTypeImplicits[CoolH2Profile](CoolH2Profile)
+  val h2JdbcTypeImplicits = new JdbcTypeImplicits[DAOH2Profile](DAOH2Profile)
   val mySQLJdbcTypeImplicits = new JdbcTypeImplicits[MySQLProfile](MySQLProfile)
   val postgresJdbcTypeImplicits = new JdbcTypeImplicits[PostgresProfile](PostgresProfile)
 }
