+++ test/scala/co/actioniq/luna/CompiledQuerySpec.scala
@@ -4,8 +4,8 @@ import java.util.concurrent.TimeUnit
 
 import co.actioniq.luna.{DBWithLogging, SlickScope}
 import co.actioniq.luna.compiled.{BoundedSeq, BoundedSeq100, InSeqDbUUIDImplicits, SlickCompiledFunctionSingleton}
-import co.actioniq.luna.dao.{CoolH2Profile, DAOUUIDQuery, DbUUID, FormValidatorMessageSeq, H2DAO, H2DAOTable, IdModel, JdbcTypeImplicits}
-import co.actioniq.luna.dao.CoolH2Profile.api._
+import co.actioniq.luna.dao.{DAOH2Profile, DAOUUIDQuery, DbUUID, FormValidatorMessageSeq, H2DAO, H2DAOTable, IdModel, JdbcTypeImplicits}
+import co.actioniq.luna.dao.DAOH2Profile.api._
 import co.actioniq.luna.logging.NoopBackend
 import org.junit.runner.RunWith
 import org.specs2.mock.Mockito
@@ -176,7 +176,7 @@ class PersistedDao(
   override protected val db: DBWithLogging,
   private val persistedSlick: TableQuery[PersistedSlickTable]
 ) extends H2DAO[PersistedSlickTable, PersistedSlick, DbUUID]
-  with DAOUUIDQuery[PersistedSlickTable, PersistedSlick, CoolH2Profile]
+  with DAOUUIDQuery[PersistedSlickTable, PersistedSlick, DAOH2Profile]
   with NoopBackend
   with InSeqDbUUIDImplicits {
   override def validateCreate(
