+++ test/scala/co/actioniq/luna/SoftDeleteSpec.scala
@@ -3,7 +3,7 @@ package co.actioniq.luna
 import java.util.concurrent.TimeUnit
 
 import co.actioniq.luna.dao.DbUUID
-import co.actioniq.luna.dao.CoolH2Profile.api._
+import co.actioniq.luna.dao.DAOH2Profile.api._
 import co.actioniq.luna.example.{Kitty, KittyDAO}
 import co.actioniq.luna.logging.NoopBackend
 import org.junit.runner.RunWith
